images: move most logic into storageUtils
